changed # of bedrooms to an integer, made sidebar bedroom/bathroom numberpickers match one another, fixed min/max price updating between the two.  WARNING: if merging this branch does not update the Prototype_master branch's mysharedprefs file (IDK if it does or not) then propty will crash because of an int to float casting error.  Tofix this, make a shared prefs editor and run editor.clear(); editor.commit(); one time to clear the old preferences.

+++ app/src/main/java/io/propty/propty/PropertyDatabaseHandler.java
@@ -98,7 +98,7 @@ public class PropertyDatabaseHandler extends SQLiteOpenHelper {
         final  String zip_string = EXTRA + "zip";
         SharedPreferences settings = context.getSharedPreferences(PREFS, 0);
 
-        float numBedrooms_val = settings.getFloat(numBedrooms_string, 1);
+        int numBedrooms_val = settings.getInt(numBedrooms_string, 1);
         float numBathrooms_val = settings.getFloat(numBathrooms_string, 1);
         int minPrice_val = settings.getInt(minPrice_string, 0);
         int maxPrice_val = settings.getInt(maxPrice_string, 0);
